+++ fabrictestbed_extensions/__init__.py
@@ -1,2 +1,3 @@
-#
-__VERSION__ = "1.3.3"
+# Please update version number here when making a release: flit, the
+# build backend we use, picks up the version from here.
+__version__ = "1.4.1"