+++ fabrictestbed_extensions/fablib/fablib.py
@@ -774,7 +774,6 @@ class FablibManager:
         self.ssh_command_line = command
 
     def set_avoid_csv(self, avoid_csv: str = ""):
-
         avoid_csv = avoid_csv.strip().strip('"').strip("'")
 
         avoid = []
@@ -1127,6 +1126,7 @@ class FablibManager:
         :return: list of random site names.
         :rtype: List[Sting]
         """
+
         # Always filter out sites in maintenance and sites that can't support any VMs
         def combined_filter_function(site):
             if filter_function is None:
@@ -1782,7 +1782,6 @@ class FablibManager:
     def show_table_jupyter(
         self, table, headers=None, title="", title_font_size="1.25em", quiet=False
     ):
-
         printable_table = pd.DataFrame(table)
 
         properties = {  # 'background-color': f'{FablibManager.FABRIC_LIGHT}',
@@ -1863,7 +1862,6 @@ class FablibManager:
         quiet=False,
         pretty_names_dict={},
     ):
-
         if output is None:
             output = self.output.lower()
 
@@ -1908,7 +1906,6 @@ class FablibManager:
         output=None,
         quiet=False,
     ):
-
         if headers is not None:
             printable_table = pd.DataFrame(table, columns=headers)
         else:
@@ -2019,7 +2016,6 @@ class FablibManager:
         filter_function=None,
         pretty_names_dict={},
     ):
-
         if filter_function:
             data = list(filter(filter_function, data))