+++ fabrictestbed_extensions/fablib/network_service.py
@@ -51,7 +51,13 @@ class NetworkService:
 
     # Type names used in fim network services
     fim_l2network_service_types = ["L2Bridge", "L2PTP", "L2STS"]
-    fim_l3network_service_types = ["FABNetv4", "FABNetv6"]
+    fim_l3network_service_types = [
+        "FABNetv4",
+        "FABNetv6",
+        "FABNetv4Ext",
+        "FABNetv6Ext",
+        "L3VPN",
+    ]
 
     @staticmethod
     def get_fim_l2network_service_types() -> List[str]:
@@ -212,8 +218,18 @@ class NetworkService:
         """
         if type == "IPv6":
             nstype = ServiceType.FABNetv6
-        else:
+        elif type == "IPv4":
             nstype = ServiceType.FABNetv4
+        elif type == "IPv4Ext":
+            nstype = ServiceType.FABNetv4Ext
+        elif type == "IPv6Ext":
+            nstype = ServiceType.FABNetv6Ext
+        elif type == "L3VPN":
+            nstype = ServiceType.L3VPN
+        else:
+            raise Exception(
+                "Invalid L3 Network Type: Allowed values [IPv4, IPv4Ext, IPv6, IPv6Ext, L3VPN]"
+            )
 
         # TODO: need a fabnet version of this
         # validate nstype and interface List
@@ -495,7 +511,6 @@ class NetworkService:
 
     @staticmethod
     def get_pretty_name_dict():
-
         return {
             "id": "ID",
             "name": "Name",
@@ -602,7 +617,7 @@ class NetworkService:
 
     def get_site(self) -> str or None:
         try:
-            return self.get_sliver().sliver.site
+            return self.get_sliver().fim_sliver.site
         except Exception as e:
             logging.warning(f"Failed to get site: {e}")
 
@@ -616,7 +631,7 @@ class NetworkService:
         :rtype: String
         """
         try:
-            return self.get_sliver().sliver.layer
+            return self.get_sliver().fim_sliver.layer
         except Exception as e:
             logging.warning(f"Failed to get layer: {e}")
             return None
@@ -629,7 +644,7 @@ class NetworkService:
         :rtype: String
         """
         try:
-            return self.get_sliver().sliver.resource_type
+            return self.get_sliver().fim_sliver.resource_type
         except Exception as e:
             logging.warning(f"Failed to get type: {e}")
             return None
@@ -674,10 +689,10 @@ class NetworkService:
         try:
             gateway = None
             if self.get_layer() == NSLayer.L3:
-                if self.get_type() == ServiceType.FABNetv6:
-                    gateway = IPv6Address(self.get_sliver().sliver.gateway.gateway)
-                elif self.get_type() == ServiceType.FABNetv4:
-                    gateway = IPv4Address(self.get_sliver().sliver.gateway.gateway)
+                if self.get_type() in [ServiceType.FABNetv6, ServiceType.FABNetv6Ext]:
+                    gateway = IPv6Address(self.get_sliver().fim_sliver.gateway.gateway)
+                elif self.get_type() in [ServiceType.FABNetv4, ServiceType.FABNetv4Ext]:
+                    gateway = IPv4Address(self.get_sliver().fim_sliver.gateway.gateway)
 
             return gateway
         except Exception as e:
@@ -720,10 +735,10 @@ class NetworkService:
         try:
             subnet = None
             if self.get_layer() == NSLayer.L3:
-                if self.get_type() == ServiceType.FABNetv6:
-                    subnet = IPv6Network(self.get_sliver().sliver.gateway.subnet)
-                elif self.get_type() == ServiceType.FABNetv4:
-                    subnet = IPv4Network(self.get_sliver().sliver.gateway.subnet)
+                if self.get_type() in [ServiceType.FABNetv6, ServiceType.FABNetv6Ext]:
+                    subnet = IPv6Network(self.get_sliver().fim_sliver.gateway.subnet)
+                elif self.get_type() in [ServiceType.FABNetv4, ServiceType.FABNetv4Ext]:
+                    subnet = IPv4Network(self.get_sliver().fim_sliver.gateway.subnet)
             return subnet
         except Exception as e:
             logging.warning(f"Failed to get subnet: {e}")