+++ fabrictestbed_extensions/fablib/node.py
@@ -175,7 +175,6 @@ class Node:
 
     @staticmethod
     def get_pretty_name_dict():
-
         return {
             "id": "ID",
             "name": "Name",
@@ -342,7 +341,6 @@ class Node:
             pretty_names_dict = {}
 
         if colors and self.get_fablib_manager().is_jupyter_notebook():
-
             table = self.get_fablib_manager().show_table(
                 data,
                 fields=fields,
@@ -1175,9 +1173,6 @@ class Node:
             f"execute node: {self.get_name()}, management_ip: {self.get_management_ip()}, command: {command}"
         )
 
-        if output_file:
-            file = open(output_file, "a")
-
         # if not quiet:
         chunking = True
 
@@ -1245,6 +1240,9 @@ class Node:
                     pkey=key,
                     sock=bastion_channel,
                 )
+
+                if output_file:
+                    file = open(output_file, "a")
 
                 # stdin, stdout, stderr = client.exec_command('echo \"' + command + '\" > /tmp/fabric_execute_script.sh; chmod +x /tmp/fabric_execute_script.sh; /tmp/fabric_execute_script.sh')
 
@@ -1339,9 +1337,6 @@ class Node:
                     rtn_stdout = b"".join(stdout_chunks).decode("utf-8")
                     rtn_stderr = b"".join(stderr_chunks).decode("utf-8")
 
-                client.close()
-                bastion_channel.close()
-
                 if self.get_fablib_manager().get_log_level() == logging.DEBUG:
                     end = time.time()
                     logging.debug(
@@ -1357,25 +1352,11 @@ class Node:
                 return rtn_stdout, rtn_stderr
                 # success, skip other tries
                 break
-            except Exception as e:
-                logging.warning(f"{e}")
-                try:
-                    client.close()
-                except:
-                    logging.debug("Exception in client.close")
-                    pass
-                try:
-                    bastion_channel.close()
-                except:
-                    logging.debug("Exception in bastion_channel.close()")
-                    pass
 
-                try:
-                    if output_file:
-                        file.close()
-                except:
-                    logging.debug("Exception in output_file close()")
-                    pass
+            except Exception as e:
+                logging.warning(
+                    f"Exception in upload_file() (attempt #{attempt} of {retry}): {e}"
+                )
 
                 if attempt + 1 == retry:
                     raise e
@@ -1389,6 +1370,29 @@ class Node:
 
                 time.sleep(retry_interval)
                 pass
+
+            # Clean-up of open connections and files.
+            finally:
+                try:
+                    client.close()
+                except Exception as e:
+                    logging.debug(f"Exception in client.close(): {e}")
+
+                try:
+                    bastion_channel.close()
+                except Exception as e:
+                    logging.debug(f"Exception in bastion_channel.close(): {e}")
+
+                try:
+                    bastion.close()
+                except Exception as e:
+                    logging.debug(f"Exception in bastion.close(): {e}")
+
+                try:
+                    if output_file:
+                        file.close()
+                except Exception as e:
+                    logging.debug(f"Exception in output_file.close(): {e}")
 
         raise Exception("ssh failed: Should not get here")
 
@@ -1496,9 +1500,6 @@ class Node:
 
                 ftp_client = client.open_sftp()
                 file_attributes = ftp_client.put(local_file_path, remote_file_path)
-                ftp_client.close()
-
-                bastion_channel.close()
 
                 if self.get_fablib_manager().get_log_level() == logging.DEBUG:
                     end = time.time()
@@ -1508,17 +1509,9 @@ class Node:
                     )
 
                 return file_attributes
+
             except Exception as e:
-                try:
-                    client.close()
-                except:
-                    logging.debug("Exception in client.close")
-                    pass
-                try:
-                    bastion_channel.close()
-                except:
-                    logging.debug("Exception in bastion_channel.close()")
-                    pass
+                logging.warning(f"Exception on upload_file() attempt #{attempt}: {e}")
 
                 if attempt + 1 == retry:
                     raise e
@@ -1530,6 +1523,27 @@ class Node:
                 # traceback.print_exc()
                 time.sleep(retry_interval)
                 pass
+
+            finally:
+                try:
+                    ftp_client.close()
+                except Exception as e:
+                    logging.debug(f"Exception in ftp_client.close(): {e}")
+
+                try:
+                    client.close()
+                except Exception as e:
+                    logging.debug(f"Exception in client.close(): {e}")
+
+                try:
+                    bastion_channel.close()
+                except Exception as e:
+                    logging.debug("Exception in bastion_channel.close(): {e}")
+
+                try:
+                    bastion.close()
+                except Exception as e:
+                    logging.debug("Exception in bastion.close(): {e}")
 
         raise Exception("scp upload failed")
 
@@ -1637,9 +1651,6 @@ class Node:
 
                 ftp_client = client.open_sftp()
                 file_attributes = ftp_client.get(remote_file_path, local_file_path)
-                ftp_client.close()
-
-                bastion_channel.close()
 
                 if self.get_fablib_manager().get_log_level() == logging.DEBUG:
                     end = time.time()
@@ -1649,28 +1660,43 @@ class Node:
                     )
 
                 return file_attributes
+
             except Exception as e:
-                try:
-                    client.close()
-                except:
-                    logging.debug("Exception in client.close")
-                    pass
-                try:
-                    bastion_channel.close()
-                except:
-                    logging.debug("Exception in bastion_channel.close()")
-                    pass
+                logging.warning(
+                    f"Exception in download_file() (attempt #{attempt} of {retry}): {e}"
+                )
 
                 if attempt + 1 == retry:
                     raise e
 
                 # Fail, try again
                 logging.warning(
-                    f"SCP upload fail. Slice: {self.get_slice().get_name()}, Node: {self.get_name()}, trying again. Exception: {e}"
+                    f"SCP download fail. Slice: {self.get_slice().get_name()}, Node: {self.get_name()}, trying again. Exception: {e}"
                 )
                 # traceback.print_exc()
                 time.sleep(retry_interval)
                 pass
+
+            finally:
+                try:
+                    ftp_client.close()
+                except Exception as e:
+                    logging.debug(f"Exception in ftp_client.close(): {e}")
+
+                try:
+                    client.close()
+                except Exception as e:
+                    logging.debug(f"Exception in client.close(): {e}")
+
+                try:
+                    bastion_channel.close()
+                except Exception as e:
+                    logging.debug(f"Exception in bastion_channel.close(): {e}")
+
+                try:
+                    bastion.close()
+                except Exception as e:
+                    logging.debug(f"Exception in bastion.close(): {e}")
 
         raise Exception("scp download failed")
 
@@ -1930,7 +1956,6 @@ class Node:
         )
 
     def ip_addr_list(self, output="json", update=False):
-
         try:
             if self.ip_addr_list_json is not None and update is False:
                 return self.ip_addr_list_json
@@ -2002,7 +2027,6 @@ class Node:
             raise e
 
     def get_ip_routes(self):
-
         """
         Get a list of routes from the node.
         """
@@ -2127,9 +2151,15 @@ class Node:
             if network is None:
                 return
             elif network.get_layer() == NSLayer.L3:
-                if network.get_type() == ServiceType.FABNetv6:
+                if network.get_type() in [
+                    ServiceType.FABNetv6,
+                    ServiceType.FABNetv6Ext,
+                ]:
                     ip_command = "sudo ip -6"
-                elif interface.get_network().get_type() == ServiceType.FABNetv4:
+                elif interface.get_network().get_type() in [
+                    ServiceType.FABNetv4,
+                    ServiceType.FABNetv4Ext,
+                ]:
                     ip_command = "sudo ip"
             else:
                 ip_command = "sudo ip"
@@ -2167,9 +2197,15 @@ class Node:
         """
         try:
             if interface.get_network().get_layer() == NSLayer.L3:
-                if interface.get_network().get_type() == ServiceType.FABNetv6:
+                if interface.get_network().get_type() in [
+                    ServiceType.FABNetv6,
+                    ServiceType.FABNetv6Ext,
+                ]:
                     ip_command = "sudo ip -6"
-                elif interface.get_network().get_type() == ServiceType.FABNetv4:
+                elif interface.get_network().get_type() in [
+                    ServiceType.FABNetv4,
+                    ServiceType.FABNetv4Ext,
+                ]:
                     ip_command = "sudo ip"
             else:
                 ip_command = "sudo ip"
@@ -2309,9 +2345,15 @@ class Node:
         try:
             gateway = None
             if interface.get_network().get_layer() == NSLayer.L3:
-                if interface.get_network().get_type() == ServiceType.FABNetv6:
+                if interface.get_network().get_type() in [
+                    ServiceType.FABNetv6,
+                    ServiceType.FABNetv6Ext,
+                ]:
                     ip_command = "sudo ip -6"
-                elif interface.get_network().get_type() == ServiceType.FABNetv4:
+                elif interface.get_network().get_type() in [
+                    ServiceType.FABNetv4,
+                    ServiceType.FABNetv4Ext,
+                ]:
                     ip_command = "sudo ip"
             else:
                 ip_command = "sudo ip"