+++ fabrictestbed_extensions/tests/component_tests.py
@@ -37,7 +37,6 @@ from .abc_test import AbcTest
 
 
 class ComponentTest(AbcTest):
-
     COMPONENT_MODELS = {
         "GPU_Tesla_T4": ComponentModelType.GPU_Tesla_T4,
         "GPU_RTX6000": ComponentModelType.GPU_RTX6000,
@@ -121,7 +120,6 @@ class ComponentTest(AbcTest):
         # print("Lease End  : {}".format(self.slice.lease_end))
 
     def config_test_GPU_RTX6000(self, node):
-
         node_name = node.name
         try:
             name = node_name
@@ -195,7 +193,6 @@ class ComponentTest(AbcTest):
                 traceback.print_exc()
 
     def config_test_NVME_P4510(self, node):
-
         node_name = node.name
         try:
             name = node_name
@@ -270,7 +267,6 @@ class ComponentTest(AbcTest):
                 traceback.print_exc()
 
     def config_test_SmartNIC_ConnectX_5(self, node):
-
         node_name = node.name
         try:
             name = node_name
@@ -344,7 +340,6 @@ class ComponentTest(AbcTest):
                 traceback.print_exc()
 
     def config_test_SmartNIC_ConnectX_6(self, node):
-
         node_name = node.name
         try:
             name = node_name
@@ -420,7 +415,6 @@ class ComponentTest(AbcTest):
                 traceback.print_exc()
 
     def config_test_SharedNIC_ConnectX_6(self, node):
-
         node_name = node.name
         try:
             name = node_name
@@ -494,7 +488,6 @@ class ComponentTest(AbcTest):
                 traceback.print_exc()
 
     def config_test_GPU_Tesla_T4(self, node):
-
         node_name = node.name
         # print("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
         # print("Node:")