+++ fabrictestbed_extensions/tests/nvme_benchmark.py
@@ -105,7 +105,6 @@ class NVMEBenchmark(AbcTest):
         pass
 
     def run_test(self):
-
         for node_name, node in self.topology.nodes.items():
             # print("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
             # print("Node:")