+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 FABRIC testbed extensions changlog
 
-##  [1.3.3] - 2022-12-01
+## [1.3.4] - 2023-01-19
+
+### Fixed
+
+- FABLIB: Better clean up of SSH connections to bastion proxy.
+
+
+## [1.3.3] - 2022-12-01
 
  
 ### Added
@@ -21,4 +28,4 @@ FABRIC testbed extensions changlog
  
 ## [1.3.2] - 2022-10-25
   
-Older and not included in change log
+Older changes are not included in change log.